Update user and application table schemas to use BIGINT for ID fields

+++ database/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id INT PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     method TEXT NOT NULL,
     rounding SMALLINT NOT NULL,
     language TEXT NOT NULL,
@@ -8,7 +8,7 @@ CREATE TABLE users (
 
 CREATE TABLE applications (
     id SERIAL PRIMARY KEY,
-    user_id INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     parameters JSONB NOT NULL,
     status TEXT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
